FE-10: Added adjustments to UI
