+++ src/components/elements/Lists/ProductList/index.tsx
@@ -1,19 +1,20 @@
-import { Col, Row } from 'antd';
 import ProductCard from '../../Cards/ProductCard';
 import useProductList from './useProductList';
+import { ProductProps } from '../../../../types/product.type';
 
 const ProductList = () => {
-  
   const { product } = useProductList();
-  
+
   return (
-    <Row gutter={[0, 24]} className="mt-10">
-      {product.map((item: any) => (
-        <Col xs={12} md={8} xl={6} xxl={4}>
-          <ProductCard product={item} />
-        </Col>
-      ))}
-    </Row>
+    <>
+      <div className="mt-10 grid grid-cols-2 md:grid-cols-3 md:gap-x-0 md:gap-y-6 xl:grid-cols-5">
+        {product.map((item: ProductProps) => (
+          <div key={item.id}>
+            <ProductCard product={item} />
+          </div>
+        ))}
+      </div>
+    </>
   );
 };
