+++ src/output.css
@@ -141,8 +141,24 @@
   bottom: 0px;
 }
 
+#root .left-\[198px\]{
+  left: 198px;
+}
+
+#root .top-20{
+  top: 5rem;
+}
+
+#root .top-\[198px\]{
+  top: 198px;
+}
+
 #root .z-0{
   z-index: 0;
+}
+
+#root .z-10{
+  z-index: 10;
 }
 
 #root .z-20{
@@ -207,12 +223,20 @@
   margin-top: 2.5rem;
 }
 
+#root .mt-5{
+  margin-top: 1.25rem;
+}
+
 #root .inline{
   display: inline;
 }
 
 #root .flex{
   display: flex;
+}
+
+#root .grid{
+  display: grid;
 }
 
 #root .\!hidden{
@@ -308,6 +332,11 @@
   flex-grow: 1;
 }
 
+#root .translate-x-1\/2{
+  --tw-translate-x: 50%;
+  transform: translate(var(--tw-translate-x), var(--tw-translate-y)) rotate(var(--tw-rotate)) skewX(var(--tw-skew-x)) skewY(var(--tw-skew-y)) scaleX(var(--tw-scale-x)) scaleY(var(--tw-scale-y));
+}
+
 @keyframes bounce{
   0%, 100%{
     transform: translateY(-25%);
@@ -322,6 +351,10 @@
 
 #root .animate-bounce{
   animation: bounce 1s infinite;
+}
+
+#root .grid-cols-2{
+  grid-template-columns: repeat(2, minmax(0, 1fr));
 }
 
 #root .items-center{
@@ -366,6 +399,11 @@
   border-radius: 0.75rem;
 }
 
+#root .bg-black{
+  --tw-bg-opacity: 1;
+  background-color: rgb(0 0 0 / var(--tw-bg-opacity));
+}
+
 #root .bg-blue-50{
   --tw-bg-opacity: 1;
   background-color: rgb(239 246 255 / var(--tw-bg-opacity));
@@ -388,6 +426,10 @@
 
 #root .bg-white\/10{
   background-color: rgb(255 255 255 / 0.1);
+}
+
+#root .bg-opacity-65{
+  --tw-bg-opacity: 0.65;
 }
 
 #root .bg-gradient-to-tr{
@@ -436,6 +478,21 @@
   padding-right: 1.25rem;
 }
 
+#root .px-7{
+  padding-left: 1.75rem;
+  padding-right: 1.75rem;
+}
+
+#root .py-10{
+  padding-top: 2.5rem;
+  padding-bottom: 2.5rem;
+}
+
+#root .py-16{
+  padding-top: 4rem;
+  padding-bottom: 4rem;
+}
+
 #root .py-3{
   padding-top: 0.75rem;
   padding-bottom: 0.75rem;
@@ -460,6 +517,11 @@
 
 #root .font-sans{
   font-family: Poppins, ui-sans-serif, system-ui, sans-serif, "Apple Color Emoji", "Segoe UI Emoji", "Segoe UI Symbol", "Noto Color Emoji";
+}
+
+#root .text-2xl{
+  font-size: 1.5rem;
+  line-height: 2rem;
 }
 
 #root .text-\[46px\]{
@@ -593,11 +655,6 @@
     margin-right: 0.75rem;
   }
 
-  #root .md\:mx-5{
-    margin-left: 1.25rem;
-    margin-right: 1.25rem;
-  }
-
   #root .md\:mb-20{
     margin-bottom: 5rem;
   }
@@ -608,6 +665,19 @@
 
   #root .md\:inline{
     display: inline;
+  }
+
+  #root .md\:grid-cols-3{
+    grid-template-columns: repeat(3, minmax(0, 1fr));
+  }
+
+  #root .md\:gap-x-0{
+    -moz-column-gap: 0px;
+         column-gap: 0px;
+  }
+
+  #root .md\:gap-y-6{
+    row-gap: 1.5rem;
   }
 
   #root .md\:py-10{
@@ -691,6 +761,10 @@
 
   #root .xl\:flex-grow{
     flex-grow: 1;
+  }
+
+  #root .xl\:grid-cols-5{
+    grid-template-columns: repeat(5, minmax(0, 1fr));
   }
 
   #root .xl\:text-base{
