+++ src/pages/ProductList/components/ListProducts/index.tsx
@@ -4,7 +4,7 @@ import ProductCard from "../../../../components/elements/Cards/ProductCard";
 const ListProducts = ({productList}: {productList: any}) => {
   return (
     <Row gutter={[0, 24]} className="mt-10">
-      {productList.map((item: any) => (
+      {productList?.map((item: any) => (
         <Col xl={6}>
           <ProductCard product={item} />
         </Col>
